+++ routers/home.js
@@ -1,0 +1,13 @@
+const express=require('express')
+const router=express.Router()
+
+
+router.get("/",(req,res)=>res.render('pages/home'))  
+router.get("/about",(req,res)=>res.render('pages/about'))  
+router.get("/contact",(req,res)=>res.render('pages/contact'))  
+router.get("/login",(req,res)=>res.render('pages/login'))  
+router.get("/signUp",(req,res)=>res.render('pages/signup'))  
+
+
+
+module.exports=router;
