+++ views/canteens/allCanteens.ejs
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>All Canteens</title>
 </head>
 <body>
-    
+    All Canteens
 </body>
 </html>
