+++ views/canteens/lipton.ejs
@@ -6,6 +6,6 @@
     <title>Document</title>
 </head>
 <body>
-    
+    lipton
 </body>
 </html>
